fix: resolve all cargo clippy warnings across workspace
Address collapsible_if, needless_borrow, needless_lifetimes,
unnecessary_cast, large_enum_variant, uninit_vec, expect_fun_call,
field_reassign_with_default, useless_conversion, and
needless_option_as_deref lints.

+++ crates/eruditio/src/formats/pml/parser.rs
@@ -92,9 +92,9 @@ pub(crate) fn pml_to_html(pml: &str) -> String {
                         pos += 1;
                     }
                     let title = &pml[title_start..pos];
-                    html.push_str(H_OPEN[level as usize]);
+                    html.push_str(H_OPEN[level]);
                     push_escape_html(&mut html, title);
-                    html.push_str(H_CLOSE[level as usize]);
+                    html.push_str(H_CLOSE[level]);
                 },
                 // Bold toggle.
                 b'b' | b'B' => {
